Refactor admin cabang attendance navigation
